Update KTGIAITHUA003 - Đếm số số 0 liên tiếp cuối n!.cpp

+++ KTGIAITHUA003 - Đếm số số 0 liên tiếp cuối n!.cpp
@@ -1,8 +1,8 @@
 #include<bits/stdc++.h>
 using namespace std;
-int dem (long long n) {
-    int count = 0;
-    for(int i = 5; i<= n ;i*= 5) {
+long long dem (long long n) {
+    long long count = 0;
+    for(long long i = 5; i<= n ;i*= 5) {
         count += n/i;
     }
     return count;
